vote_app/requirements.txt
flask
boto3
watchtower==3.0.1
python-dotenv